# Conflicts:
#	yggdrash-common/src/main/java/io/yggdrash/common/config/Constants.java

+++ yggdrash-core/src/main/java/io/yggdrash/core/store/PeerStore.java
@@ -16,62 +16,71 @@
 
 package io.yggdrash.core.store;
 
+import io.yggdrash.common.crypto.HashUtil;
 import io.yggdrash.common.store.datasource.DbSource;
+import io.yggdrash.common.utils.ByteUtil;
+import io.yggdrash.common.utils.SerializationUtil;
 import io.yggdrash.contract.core.store.ReadWriterStore;
+import io.yggdrash.core.exception.NonExistObjectException;
 import io.yggdrash.core.p2p.Peer;
 import io.yggdrash.core.p2p.PeerId;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.spongycastle.util.encoders.Hex;
 
-import java.io.IOException;
-import java.util.HashMap;
+import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 
 public class PeerStore implements ReadWriterStore<PeerId, Peer> {
 
-    private static final Logger log = LoggerFactory.getLogger(PeerStore.class);
+    private static final byte[] TOTAL_SIZE = SerializationUtil.serializeString("TOTAL_SIZE");
+
     private final DbSource<byte[], byte[]> db;
-    private final Map<PeerId, Peer> peers = new HashMap<>();
+    private long peerSize;
 
     PeerStore(DbSource<byte[], byte[]> dbSource) {
         this.db = dbSource.init();
+        this.peerSize = loadPeerSize();
     }
 
     @Override
     public void put(PeerId key, Peer value) {
-        peers.put(key, value);
+        if (!contains(key)) {
+            peerSize++;
+            byte[] indexKey = getIndexKey(peerSize);
+            // store peer index
+            db.put(indexKey, key.getBytes());
+            db.put(TOTAL_SIZE, ByteUtil.longToBytes(peerSize));
+        }
         db.put(key.getBytes(), value.toString().getBytes());
-    }
-
-    private void put(Peer value) {
-        put(value.getPeerId(), value);
     }
 
     @Override
     public Peer get(PeerId key) {
-        byte[] foundedValue = db.get(key.getBytes());
+        return get(key.getBytes());
+    }
+
+    private Peer get(byte[] key) {
+        byte[] foundedValue = db.get(key);
         if (foundedValue != null) {
             return Peer.valueOf(foundedValue);
         }
-        return peers.get(key);
+        throw new NonExistObjectException(Hex.toHexString(key));
     }
 
     @Override
     public boolean contains(PeerId key) {
-        if (db.get(key.getBytes()) != null) {
-            return true;
-        }
-        return peers.containsKey(key);
+        return db.get(key.getBytes()) != null;
     }
 
     public void overwrite(List<Peer> peerList) {
-
-        peers.keySet().stream().map(PeerId::getBytes).forEach(db::delete);
-        peers.keySet().removeIf(peers::containsKey);
-
-        peerList.forEach(this::put);
+        // remove all
+        for (int i = 1; i <= peerSize; i++) {
+            byte[] indexKey = getIndexKey(i);
+            byte[] key = db.get(indexKey);
+            db.delete(key);
+        }
+        peerSize = 0L;
+        // put all
+        peerList.forEach(peer -> put(peer.getPeerId(), peer));
     }
 
     public void close() {
@@ -79,30 +88,50 @@ public class PeerStore implements ReadWriterStore<PeerId, Peer> {
     }
 
     public void remove(PeerId key) {
-        peers.remove(key);
+        byte[] foundedValue = db.get(key.getBytes());
+        if (foundedValue == null) {
+            return;
+        }
+        byte[] indexKey = getIndexKey(peerSize);
+        db.delete(indexKey);
         db.delete(key.getBytes());
+        peerSize--;
+        db.put(TOTAL_SIZE, ByteUtil.longToBytes(peerSize));
     }
 
     public List<String> getAll() {
-        try {
-            if (!db.getAll().isEmpty()) {
-                return db.getAll().stream().map(Peer::valueOf)
-                        .map(Peer::toString).collect(Collectors.toList());
-            }
-        } catch (IOException e) {
-            log.error(e.getMessage());
+        List<String> list = new ArrayList<>();
+        if (peerSize == 0) {
+            return list;
         }
-        return peers.values().stream().map(Peer::toString).collect(Collectors.toList());
+
+        for (int i = 1; i <= peerSize; i++) {
+            byte[] indexKey = getIndexKey(i);
+            byte[] key = db.get(indexKey);
+            Peer peer = get(key);
+            list.add(peer.getYnodeUri());
+        }
+        return list;
     }
 
-    public int size() {
-        try {
-            if (!db.getAll().isEmpty()) {
-                return db.getAll().size();
-            }
-        } catch (IOException e) {
-            log.error(e.getMessage());
+    public long size() {
+        return peerSize;
+    }
+
+    private long loadPeerSize() {
+        if (peerSize > 0L) {
+            return peerSize;
         }
-        return peers.size();
+        // loading db is just first
+        byte[] sizeByte = db.get(TOTAL_SIZE);
+        if (sizeByte != null) {
+            return ByteUtil.byteArrayToLong(sizeByte);
+        }
+        return peerSize;
+    }
+
+    private byte[] getIndexKey(long index) {
+        String blockIndexKey = "PEER_INDEX_" + index;
+        return HashUtil.sha3(blockIndexKey.getBytes());
     }
 }
